entry_std: Disable OPTEE_MSG_CMD_CANCEL SMC
The new task feature that is introduced in the next patch is not
compatible with the way the cancel SMC is currently handled. Disable
that SMC completely for now so that the new task feature can be
committed. The SMC will be reinstated by a later patch of this series.