task: Implement new feature "tasks"
Execute the TA entry points OpenSession, InvokeCommand, and CloseSession
within the context of a "task". The primary purpose of a task is to
manage its own cancellation status. This changes the previous model,
where cancellations were managed per-session, to a model where
cancellations are managed per-task. This allows individual tasks to be
cancelled without affecting other tasks that are associated with the
same session.

The main part of this patch is the implementation of the feature in
task.c and task.h.  Most of the other files touched by this patch are
just due changed function names and no actual logic is changed in those
files.

The important changes are in entry_std.c and tee_svc: In these files,
new tasks are created and destroyed before and after calls to
OpenSession, InvokeCommand, and CloseSession.